requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Performance Optimizations
slowapi==0.1.9
aiofiles==23.2.1
cachetools==5.3.2

# Frontend
streamlit==1.28.1
pandas==2.1.3

# Configuration & Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
requests-mock==1.11.0
httpx==0.25.2

# Development & Monitoring
psutil==5.9.6
memory-profiler==0.61.0